small sidplayfp update / last main branch commits; atariaudio & v2m now generates midi note

+++ libs/sid/libsidplayfp/src/builders/residfp-builder/residfp/resample/SincResampler.h
@@ -1,7 +1,7 @@
 /*
  * This file is part of libsidplayfp, a SID player engine.
  *
- * Copyright 2011-2013 Leandro Nini <drfiemost@users.sourceforge.net>
+ * Copyright 2011-2024 Leandro Nini <drfiemost@users.sourceforge.net>
  * Copyright 2007-2010 Antti Lankila
  * Copyright 2004 Dag Lem <resid@nimrod.no>
  *
@@ -74,7 +74,7 @@ private:
 
     int outputValue;
 
-    short sample[RINGSIZE * 2];
+    int sample[RINGSIZE * 2];
 
 private:
     int fir(int subcycle);
